# dashboard-stub/api/swagger.yaml
...
swagger: "2.0"
info:
  version: "0.0.1"
  title: "Techorrect Test Dashboard"
host: "localhost"
basePath: "/dashboard"
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /projects:
    get:
      description: "Gets the projects that the logged in user is authorized to view.\n"
      operationId: "projectsGET"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - name: "name"
        in: "query"
        required: false
        type: "string"
      responses:
        200:
          description: "Successful response"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Projects"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
    post:
      description: "Creates a project\n"
      operationId: "projectsPOST"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - in: "body"
        name: "project"
        required: true
        schema:
          $ref: "#/definitions/Projects"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/Projects"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
    options:
      description: "Options operation\n"
      operationId: "projectsOPTIONS"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Default"
  /projects/{projectId}:
    get:
      description: "Gets one project\n"
      operationId: "projectsProjectIdGET"
      parameters:
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful response"
          schema:
            $ref: "#/definitions/Projects"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
    put:
      description: "Updates a project\n"
      operationId: "projectsProjectIdPUT"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "project"
        required: true
        schema:
          $ref: "#/definitions/Projects"
      responses:
        200:
          description: "Successful response"
          schema:
            $ref: "#/definitions/Projects"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
    delete:
      description: "Deletes a project\n"
      operationId: "projectsProjectIdDELETE"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      responses:
        204:
          description: "Successfully deleted"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
    options:
      description: "Options operation\n"
      operationId: "projectsProjectIdOPTIONS"
      parameters:
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Default"
  /projects/{projectId}/suites:
    get:
      description: "Gets the suites that have run for a given project\n"
      operationId: "projectsProjectIdSuitesGET"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      - name: "prevRuns"
        in: "query"
        required: false
        type: "integer"
        default: 25
      - name: "name"
        in: "query"
        required: false
        type: "string"
      responses:
        200:
          description: "Successful response"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Suites"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
    post:
      description: "Creates a suite for a given project\n"
      operationId: "projectsProjectIdSuitesPOST"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "suite"
        required: true
        schema:
          $ref: "#/definitions/Suites"
      responses:
        200:
          description: "OK"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
    options:
      description: "Options operation\n"
      operationId: "projectsProjectIdSuitesOPTIONS"
      parameters:
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Default"
  /projects/{projectId}/suites/{suiteId}:
    get:
      description: "Gets one suite\n"
      operationId: "projectsProjectIdSuitesSuiteIdGET"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      - name: "suiteId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful response"
          schema:
            $ref: "#/definitions/Suites"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
    put:
      description: "Updates a suite\n"
      operationId: "projectsProjectIdSuitesSuiteIdPUT"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      - name: "suiteId"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "suite"
        required: true
        schema:
          $ref: "#/definitions/Suites"
      responses:
        200:
          description: "Successful response"
          schema:
            $ref: "#/definitions/Suites"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
    delete:
      description: "Deletes a suite\n"
      operationId: "projectsProjectIdSuitesSuiteIdDELETE"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      - name: "suiteId"
        in: "path"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
    options:
      description: "Options operation\n"
      operationId: "projectsProjectIdSuitesSuiteIdOPTIONS"
      parameters:
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      - name: "suiteId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Default"
  /projects/{projectId}/suites/{suiteId}/{suiteRunId}:
    get:
      description: "Gets the test runs that ran as part of a particular suite run.\n"
      operationId: "projectsProjectIdSuitesSuiteIdSuiteRunIdGET"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      - name: "suiteId"
        in: "path"
        required: true
        type: "string"
      - name: "suiteRunId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful response"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/TestRuns"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
  /projects/{projectId}/suites/{suiteId}/tests:
    get:
      description: "Gets the tests that have run for a given suite\n"
      operationId: "projectsProjectIdSuitesSuiteIdTestsGET"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      - name: "suiteId"
        in: "path"
        required: true
        type: "string"
      - name: "name"
        in: "query"
        required: false
        type: "string"
      responses:
        200:
          description: "Successful response"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Tests"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
    post:
      description: "Creates a test that runs for a given suite\n"
      operationId: "projectsProjectIdSuitesSuiteIdTestsPOST"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      - name: "suiteId"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "test"
        required: true
        schema:
          $ref: "#/definitions/Tests"
      responses:
        200:
          description: "Successful response"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Tests"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
    options:
      description: "Options operation\n"
      operationId: "projectsProjectIdSuitesSuiteIdTestsOPTIONS"
      parameters:
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      - name: "suiteId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Default"
  /projects/{projectId}/suites/{suiteId}/tests/{testId}:
    get:
      description: "Gets one test\n"
      operationId: "projectsProjectIdSuitesSuiteIdTestsTestIdGET"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      - name: "suiteId"
        in: "path"
        required: true
        type: "string"
      - name: "testId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful response"
          schema:
            $ref: "#/definitions/Tests"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
    put:
      description: "Updates a test\n"
      operationId: "projectsProjectIdSuitesSuiteIdTestsTestIdPUT"
      parameters:
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      - name: "suiteId"
        in: "path"
        required: true
        type: "string"
      - name: "testId"
        in: "path"
        required: true
        type: "string"
      - in: "body"
        name: "test"
        required: true
        schema:
          $ref: "#/definitions/Tests"
      responses:
        200:
          description: "Successful response"
          schema:
            $ref: "#/definitions/Tests"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
    delete:
      description: "Deletes a test\n"
      operationId: "projectsProjectIdSuitesSuiteIdTestsTestIdDELETE"
      parameters:
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      - name: "suiteId"
        in: "path"
        required: true
        type: "string"
      - name: "testId"
        in: "path"
        required: true
        type: "string"
      responses:
        204:
          description: "Successfully deleted"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
    options:
      description: "Options operation\n"
      operationId: "projectsProjectIdSuitesSuiteIdTestsTestIdOPTIONS"
      parameters:
      - name: "projectId"
        in: "path"
        required: true
        type: "string"
      - name: "suiteId"
        in: "path"
        required: true
        type: "string"
      - name: "testId"
        in: "path"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Default"
...
  /users:
    post:
      description: "Creates a user and authorizes that user for a list of projects.\n"
      operationId: "usersPOST"
      parameters:
      - name: "x-access-token"
        in: "header"
        required: true
        type: "string"
      - in: "body"
        name: "user"
        required: true
        schema:
          $ref: "#/definitions/Users"
      responses:
        201:
          description: "Successful response of created user object excluding password."
          schema:
            $ref: "#/definitions/Users"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "Default"
    options:
      description: "Options operation\n"
      operationId: "usersOPTIONS"
      parameters: []
      responses:
        200:
          description: "OK"
      x-swagger-router-controller: "Default"
...
definitions:
  Projects:
    type: "object"
    properties:
      projectId:
        type: "string"
        example: "project1"
      name:
        type: "string"
        example: "Project name 1"
      healthySuites:
        type: "integer"
        example: 30
      unhealthySuites:
        type: "integer"
        example: 2
      healtyFailedLevel:
        type: "integer"
        example: 5
      healthySkippedLevel:
        type: "integer"
        example: 5
      healthyDisabledLevel:
        type: "integer"
        example: 10
    example:
      unhealthySuites: 2
      healtyFailedLevel: 5
      healthyDisabledLevel: 10
      name: "Project name 1"
      projectId: "project1"
      healthySkippedLevel: 5
      healthySuites: 30
  Suites:
    type: "object"
    properties:
      suiteId:
        type: "string"
        example: "suite1"
      name:
        type: "string"
        example: "Suite name 1"
      passingTests:
        type: "integer"
        example: 100
      failingTests:
        type: "integer"
        example: 4
      skippedTests:
        type: "integer"
        example: 3
      disabledTests:
        type: "integer"
        example: 3
      previousSuiteIds:
        type: "array"
        items:
          type: "string"
          example: ""
    example:
      skippedTests: 3
      suiteId: "suite1"
      disabledTests: 3
      failingTests: 4
      passingTests: 100
      name: "Suite name 1"
      previousSuiteIds:
      - ""
      - ""
...
  Users:
    type: "object"
    properties:
      username:
        type: "string"
      password:
        type: "string"
      admin:
        type: "boolean"
      authorizedProjects:
        type: "string"
    example:
      password: "password"
      admin: true
      username: "username"
      authorizedProjects: "authorizedProjects"